src/api/valet-api/pyproject.toml
[project]
name = "valet-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.3",
    "azure-identity>=1.25.2",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.133.0",
    "fastapi-cli>=0.0.24",
    "python-dotenv>=1.2.1",
]
